TypeScript Node Implementation - added TypeScript to Server - fixed minor bugs in task-manager
