+++ app/index.html
@@ -1,27 +1,32 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <title>
+      Ariki's TODO-list
+    </title>
+    <link
+      rel="stylesheet"
+      href="../assets/stylesheets/task.css"
+      type="text/css"
+    />
+  </head>
 
-<head>
-  <meta charset="utf-8" />
-  <title>
-    Ariki's TODO-list
-  </title>
-  <link rel="stylesheet" href="../assets/stylesheets/task.css" type="text/css" />
-</head>
-
-<body>
-  <h1 class="title">
-    Ariki's TODO-list
-  </h1>
-  <div class="container">
-    <div class="taskList"></div>
-    <div class="submission">
-      <small> What do you need to do?</small>
-      <input class="newTaskInput" />
+  <body>
+    <h1 class="title">
+      Ariki's TODO-list
+    </h1>
+    <div class="container">
+      <div class="taskList"></div>
+      <div class="submission">
+        <small> What do you need to do?</small>
+        <input class="newTaskInput" />
+      </div>
+      <button class="submitButton">Add Item</button>
     </div>
-    <button class="submitButton">Add Item</button>
-  </div>
-  <script type="text/javascript" src="../dist/task-manager.js"></script>
-</body>
-
+    <script
+      type="text/javascript"
+      src="../dist/client/task-manager.js"
+    ></script>
+  </body>
 </html>
