+++ package.json
@@ -4,7 +4,7 @@
   "description": "My First Project",
   "main": "server/server.js",
   "scripts": {
-    "start": "node server/server.js",
+    "start": "node dist/server/server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -21,6 +21,7 @@
     "typescript": "^3.9.2"
   },
   "dependencies": {
+    "@types/express": "^4.17.6",
     "body-parser": "^1.19.0",
     "express": "^4.17.1"
   }
